feat: Implement document edge detection for camera and image processing
- Added DocumentEdgeDetector for detecting document edges in images and camera frames.
- Integrated live document edge overlay in CameraScreen to provide real-time feedback.
- Updated ScanFilterService to utilize the new edge detection algorithm for improved document cropping.
- Refactored image processing logic in ImageEditorService and ScanFilterService for better performance and clarity.
- Enhanced error handling and user experience during image capture and processing.

+++ lib/core/services/image_editor_service.dart
@@ -214,11 +214,13 @@ class ImageEditorService {
     img.Point toPixel(Offset normalized) {
       final px = (normalized.dx * image.width)
           .round()
-          .clamp(0, image.width - 1);
+          .clamp(0, image.width - 1)
+          .toInt();
 
       final py = (normalized.dy * image.height)
           .round()
-          .clamp(0, image.height - 1);
+          .clamp(0, image.height - 1)
+          .toInt();
 
       return img.Point(px, py);
     }
